requirements.txt
# متطلبات مشروع twitter-ai-bot
tweepy==4.14.0
feedparser==6.0.11
deep-translator==1.11.4
requests==2.31.0
python-dotenv==1.0.0
pytz==2024.1
beautifulsoup4==4.12.3
lxml==5.1.0
